requirements.txt
selene~=2.0.0b16
pytest~=7.2.0
allure-pytest~=2.12.0
allure-python-commons~=2.12.0